backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9

# Database & Supabase
supabase>=2.12.0
pgvector>=0.3.5
asyncpg>=0.31.0
sqlalchemy>=2.0.35

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI & RAG
langchain>=0.3.2
langchain-community>=0.3.2
langchain-groq>=0.2.0
langchain-nomic>=0.1.2
langchain-text-splitters>=0.3.0
nomic>=3.1.2
tiktoken>=0.7.0
# Explicitly pin numpy to get py3.13 wheels
numpy>=2.0.0

# Document Processing & OCR
pymupdf>=1.24.11
pytesseract>=0.3.13
Pillow>=10.4.0
beautifulsoup4>=4.12.3

# Voice
elevenlabs==1.9.0
python-socketio==5.11.4

# Config & Utils
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
cloudinary>=1.41.0
email-validator>=2.2.0

# Dev
httpx==0.27.2
pytest==8.3.3
